Split braintree-web into modules
